feat(doctor): add helm doctor command
- move all prereq checks under the doctor umbrella
