+++ action/install.go
@@ -13,13 +13,13 @@ import (
 	"github.com/helm/helm/log"
 )
 
-// The order in which manifests should be installed, by Kind.
+// InstallOrder defines the order in which manifests should be installed, by Kind.
 //
 // Anything not on the list will be installed after the last listed item, in
 // an indeterminate order.
 var InstallOrder = []string{"Namespace", "Secret", "PersistentVolume", "ServiceAccount", "Service", "Pod", "ReplicationController", "DaemonSet", "Ingress", "Job"}
 
-// The order in which manifests are uninstalled.
+// UninstallOrder defines the order in which manifests are uninstalled.
 //
 // Unknown manifest types (those not explicitly referenced in this list) will
 // be uninstalled before any of these, since we know that none of the core
@@ -32,7 +32,6 @@ var UninstallOrder = []string{"Service", "Pod", "ReplicationController", "Daemon
 //
 // During install, manifests are sent to Kubernetes in the ordered specified by InstallOrder.
 func Install(chartName, home, namespace string, force bool, dryRun bool) {
-
 	ochart := chartName
 	r := mustConfig(home).Repos
 	table, chartName := r.RepoChart(chartName)
@@ -64,6 +63,8 @@ func Install(chartName, home, namespace string, force bool, dryRun bool) {
 			log.Die("Stopping install. Re-run with --force to install anyway.")
 		}
 	}
+
+	CheckKubePrereqs()
 
 	msg := "Running `kubectl create -f` ..."
 	if dryRun {
