+++ action/update.go
@@ -1,14 +1,10 @@
 package action
 
 import (
-	"io/ioutil"
-	"os"
-	"os/exec"
 	"path/filepath"
 
 	"github.com/Masterminds/semver"
 
-	"github.com/helm/helm/config"
 	"github.com/helm/helm/log"
 	"github.com/helm/helm/release"
 )
@@ -20,9 +16,7 @@ func Update(home string) {
 		log.Die("Could not generate absolute path for %q: %s", home, err)
 	}
 
-	// Basically, install if this is the first run.
-	ensurePrereqs()
-	ensureHome(home)
+	CheckLocalPrereqs(home)
 
 	rc := mustConfig(home).Repos
 	if err := rc.UpdateAll(); err != nil {
@@ -63,44 +57,4 @@ func CheckLatest(version string) {
 		}
 	}
 
-}
-
-// ensurePrereqs verifies that Git and Kubectl are both available.
-func ensurePrereqs() {
-	if _, err := exec.LookPath("git"); err != nil {
-		log.Die("Could not find 'git' on $PATH: %s", err)
-	}
-	if _, err := exec.LookPath("kubectl"); err != nil {
-		log.Die("Could not find 'kubectl' on $PATH: %s", err)
-	}
-}
-
-// ensureHome ensures that a HELM_HOME exists.
-func ensureHome(home string) {
-
-	must := []string{home, filepath.Join(home, CachePath), filepath.Join(home, WorkspacePath), filepath.Join(home, CacheChartPath)}
-
-	for _, p := range must {
-		if fi, err := os.Stat(p); err != nil {
-			log.Debug("Creating %s", p)
-			if err := os.MkdirAll(p, 0755); err != nil {
-				log.Die("Could not create %q: %s", p, err)
-			}
-		} else if !fi.IsDir() {
-			log.Die("%s must be a directory.", home)
-		}
-	}
-
-	refi := filepath.Join(home, Configfile)
-	if _, err := os.Stat(refi); err != nil {
-		log.Info("Creating %s", refi)
-		// Attempt to create a Repos.yaml
-		if err := ioutil.WriteFile(refi, []byte(config.DefaultConfigfile), 0755); err != nil {
-			log.Die("Could not create %s: %s", refi, err)
-		}
-	}
-
-	if err := os.Chdir(home); err != nil {
-		log.Die("Could not change to directory %q: %s", home, err)
-	}
 }
